+++ gradle-jdks-setup/src/main/java/com/palantir/gradle/jdks/setup/CertificateUtils.java
@@ -17,40 +17,14 @@
 package com.palantir.gradle.jdks.setup;
 
 import java.io.ByteArrayInputStream;
-import java.math.BigInteger;
-import java.nio.charset.StandardCharsets;
-import java.security.cert.Certificate;
-import java.security.cert.CertificateEncodingException;
 import java.security.cert.CertificateException;
 import java.security.cert.CertificateFactory;
 import java.security.cert.X509Certificate;
 import java.util.ArrayList;
-import java.util.Base64;
 import java.util.Collections;
 import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import java.util.stream.Stream;
 
 public final class CertificateUtils {
-    private static final BigInteger PALANTIR_3RD_GEN_SERIAL = new BigInteger("18126334688741185161");
-    private static final String PALANTIR_3RD_GEN_CERTIFICATE = "Palantir3rdGenRootCa";
-
-    public static Optional<AliasContentCert> selectPalantirCertificate(byte[] multipleCertificateBytes) {
-        return selectCertificates(
-                        multipleCertificateBytes,
-                        Map.of(PALANTIR_3RD_GEN_SERIAL.toString(), PALANTIR_3RD_GEN_CERTIFICATE))
-                .findFirst();
-    }
-
-    private static Stream<AliasContentCert> selectCertificates(
-            byte[] multipleCertificateBytes, Map<String, String> certSerialNumbersToAliases) {
-        return parseCerts(multipleCertificateBytes).stream()
-                .filter(cert -> certSerialNumbersToAliases.containsKey(
-                        cert.getSerialNumber().toString()))
-                .map(cert -> new AliasContentCert(
-                        certSerialNumbersToAliases.get(cert.getSerialNumber().toString()), encodeCertificate(cert)));
-    }
 
     public static List<X509Certificate> parseCerts(byte[] multipleCertificateBytes) {
         CertificateFactory certificateFactory;
@@ -85,18 +59,5 @@ public final class CertificateUtils {
         }
 
         return Collections.unmodifiableList(certs);
-    }
-
-    private static String encodeCertificate(Certificate palantirCert) {
-        Base64.Encoder encoder = Base64.getMimeEncoder(64, "\n".getBytes(StandardCharsets.UTF_8));
-        try {
-            return String.join(
-                    "\n",
-                    "-----BEGIN CERTIFICATE-----",
-                    encoder.encodeToString(palantirCert.getEncoded()),
-                    "-----END CERTIFICATE-----");
-        } catch (CertificateEncodingException e) {
-            throw new RuntimeException("Could not convert Palantir cert back to regular", e);
-        }
     }
 }
